Add the Apache License, Version 2.0
As of the 5.0.0 release, the UnboundID LDAP SDK for Java is
available under the terms of the Apache License, Version 2.0.  This
is a very permissive OSI-approved open source license that should be
suitable for just about any use case.  We recommend choosing this
license for all new uses of the LDAP SDK, but it remains available
under the terms of the former GPLv2, LGPLv2.1, and UnboundID LDAP
SDK Free Use License options.