Add an ldif-diff tool
Added an ldif-diff tool that can be used to compare the contents of
two LDIF files.  The differences will be written as the set of LDIF
add, delete, and modify change records needed to convert the set of
entries in the source LDIF file into the set of entries contained
in the target LDIF file.