Improve manage-certificates alt name validation
Updated the manage-certificates tool to improve the validation that
it performs for DNS names and email addresses provided for use in
the subject alternative name or issue alternative name extensions.
These fields are declared with an IA5 string syntax, which means
that their values are only allowed to contain ASCII characters.
Previously, attempts to use a value with a non-ASCII character would
fail with an internal error when attempting to create the extension.
The tool will now report a more user-friendly error message.