Update a ds-pwp-state-json field name
Updated support for the ds-pwp-state-json virtual attribute to
change the name of the field used to hold the integer representation
of the type for password policy state account usability errors,
warnings, and notices.  The field name used to be "type-value", but
that could cause some confusion as to what it actually represented.
It is now the less confusing "type-id".  This has not been
included in any release, so it does not affect backward
compatibility.