Add an option to allow empty schema element DESCs
Updated the code used to parse schema elements to allow the DESC
component to be an empty string.  While empty quoted strings are not
allowed in schema elements as per RFC 4512 section 4.1, some servers
permit the DESC element to violate this restriction, and it may
therefore be necessary to parse schema elements with empty
descriptions.  Empty descriptions are still not allowed by default,
but this can be now overridden with a code change or by setting a
system property.