Add retention support to the gen profile task
Updated the generate server profile task to include file retention
support so that previous profile zip files may be deleted based on
the count or age of existing files.

The Javadoc documentation for the profile root property has also
been updated to indicate that its value can be a directory when
the generated server profile will be packaged into a zip file, in
which case the server will generate the name for the zip file.