Add mechanism-specific SASL help for tools
Updated command-line tools so that if the --help-sasl argument is
used in conjunction with a --saslOption argument that provides a
"mech" option that specifies a mechanism name, then only the options
for that mechanism will be displayed.  If no mechanism is specified,
or if the specified mechanism is not recognized, then information
will be displayed for all available mechanisms.