Return subentries by objectClass in in-memory DS
Updated the in-memory directory server so that it can return entries
with the ldapSubEntry object class if the filter includes an
"(objectClass=ldapSubEntry)" component.  The server may also return
entries without the ldapSubEntry object class if the filter matches
them.  For example, to return all entries that do and do not contain
the ldapSubEntry object class, you may now search with a filter of
"(|(objectClass=*)(objectClass=ldapSubEntry))".