Fix an issue with usage args in properties files
Fixed an issue in the command-line tool framework that could arise
when setting the value of a usage argument through a properties
file.  In some cases, the tool could exit with an error if the usage
argument was set through a properties file even though it would have
succeeded if that same argument had been provided directly on the
command line.