Add VersionMonitorEntry.getRevisionID
Updated the support for the version monitor entry in the Ping
Identity Directory Server to provide a getRevisionID method that can
be used to retrieve a string that identifies the source revision
from which the server was built.  The former getRevisionNumber
method has been deprecated, as the server may not have been checked
out from a version control system that uses numeric revision IDs.

Also, fixed an issue in debug logging that could prevent a debug
message including an exception with an underlying cause from being
parsed as a valid JSON object.