Catch errors in message property files early
Updated the build process so that the code used to generate source
code from message properties files will catch a lot of common
errors very early in the build process.  In some cases, problems in
message property files could cause build failures that are not
obviously the result of a poorly formatted message.