Improved detection of extra spaces in messages
Improved validation for message format strings in properties files
to better detect messages with extra spaces.  Although a recently
added unit test checked for this, a bug in that test caused it to
only examine the first line of each message, while skipping
continued lines (which should start with two spaces).  This check
is now also performed at the beginning of the build when the message
property files are processed to generate source code, so any such
problems are now detected much earlier in the build process (and are
caught even when not running unit tests).  Finally, several
instances of extra spaces found by these checks have been removed
from message format strings.