Update in-memory DS tool to validate custom schema
Updated the command-line tool that may be used to launch an instance
of the in-memory directory server so that it will validate any
schema definitions provided via the --useSchemaFile argument and
report error or warning messages for any problems that it identifies
with those definitions.  Note that other than potentially displaying
additional messages at startup, this will not actually change the
behavior of the server in any way.  It will still attempt to use
those schema definitions to the best of its ability, but it will now
be easier to detect problems that could affect their use whereas
such issues had previously been silently ignored.  These messages
may be suppressed using the new --doNotValidateSchemaDefinitions if
they are not wanted.