Improve TLS trust for Ping Identity servers
Added a new TopologyRegistryTrustManager class that can read the
configuration file for a Ping Identity Directory Server instance to
obtain information about certificates for that instance and other
instances in the topology.

Updated the LDAP command-line tool framework so that if it detects
that the tool is running as part of a Ping Identity Directory
Server installation, and if no trust-related arguments are provided
for that tool, then it will include the server's default trust
store and topology registry in the set of sources that that it will
consider (which already included the JVM-default trust manager)
before resorting to interactively prompting the user about whether
to trust the presented certificate chain.  In most cases, this will
allow these tools to trust the local instance, and other instances
in the same topology, for secure communication without the need for
providing trust-related arguments on the command line.