Add a requireMatch argument to ldapsearch
Added a "--requireMatch" argument to the ldapsearch command-line
tool.  If this argument is provided and the search operation
completes successfully but does not return any matching entries,
then the tool will complete with an exit code of 94
(noResultsReturned), rather than the usual exit code of 0 (success).
This argument only affects the tool exit code; it will not have any
visible effect on the output.