Change the ldif-diff output order
Updated the ldif-diff tool so that it writes add records first,
then modifies, and then deletes.  It previously wrote modifies
first, then deletes, and then adds, but this may not be the
optimal order when there are modifies that reference added
entries (e.g., if a new user was created an added to a
group).  If the server enforces referential integrity for modify
operations, then it may reject attempts to reference an entry that
does not exist.