Fix JVM-default trust manager cross-cert handling
Fixed an issue in which the JVM-default trust manager did not
always properly handle cross-signed certificates.  If the presented
certificate chain included an issuer certificate that was outside of
its validity window, then that chain would have always been
rejected.  The trust manager will now check to see whether it is
possible to construct an alternate path to a different trusted root
that is currently valid, and if so, then the presented chain will
be considered trusted.