List and export certs in JVM-default trust store
Updated the manage-certificates tool make it easier to list and
export certificates from the JMV's default trust store.  The new
--use-jvm-default-trust-store argument can be used as an alternative
to the existing --keystore argument to indicate that it should use
the JVM's default set of trusted issuer certificates without the
need to provide the path to the appropriate cacerts file.