Add legacy support to ldifmodify
Updated the ldifmodify tool to add legacy support for a usage in
the Ping Identity Directory Server.  This mechanism uses provided
LDIF readers to obtain the source entries and changes to apply, an
LDIF writer to which the output should be written, and a list to
which error messages should be added.