Swap LDIF tool compression+encryption order
Fixed an issue with the ldifsearch, ldifmodify, and ldif-diff tools
that could arise if they were run in a manner that would case the
output to be both compressed and encrypted.  The data should have
been compressed before it was encrypted, but instead the compression
was applied after encryption.  This makes the compression
ineffective, and it also makes the output less compatible with tools
that rightfully expect compression to be applied before encryption.