Allow reading ldapcompare AV from file
Updated the ldapcompare tool to allow reading the raw assertion
value from a file.  If this option is used, then the attribute name
or OID should be followed by a single colon, a less-than sign, and
the path to the file from which the value should be read.  Note that
if this option is used, then the exact bytes of the specified file,
including any line breaks, will be used as the assertion value.