Write non-LDIF ldifsearch output as LDIF comments
Updated ldifsearch so that all non-LDIF output is written as LDIF
comments so that output will not interfere with the ability to parse
the remaining output as LDIF.

Also, hid the size limit and time limit arguments because support
for them had not been implemented.  In retrospect, it really doesn't
make sense to support a time limit option, and trying to maintain a
size limit can be problematic when processing multiple searches.