Use ^ as a continuation character on Windows
Updated the LDAP SDK so that when displaying a command that is split
across multiple lines, it will use an appropriate continuation
character for the underlying platform.  It previously always used
the backslash character (also known as a reverse solidus) for this
purpose, which is appropriate for Linux and UNIX-based systems, but
it now uses the caret (or circumflex accent) character on Windows.