Add a get recent login history control
Added support for a new get recent login history control that will
be supported in an upcoming release of the Ping Identity Directory
Server.  The request control may be included in a bind request to
indicate that a successful bind result should include information
about other recent successful and failed authentication attempts for
that account.