Add PWP state JSON recent login history support
Updated support for the Ping Identity Directory Server's
ds-pwp-state-json virtual attribute to handle additional fields
related to a user's recent login history.  This includes fields that
can be used to retrieve the history, the maximum number of
successful and failed login attempts the server should maintain, and
the maximum age in seconds of successful and failed login attempts
that the server should maintain.