Add recent login history control support to tools
Updated ldapsearch and ldapmodify to add support for the get recent
login history request control.

Updated the control argument to make it easier to use the get recent
login history request control in tools that support the argument
type by specifying a human-readable name (like
"get-recent-login-history") as an alternative to the OID for the
control.

Updated the ResultUtils class to provide support for formatting the
contents get recent login history response controls.