Update CSD support for log head and tail sizes
Updated support for the collect-support-data functionality offered
by the Ping Identity Directory Server to include the ability to
specify the amount of data to capture from the beginning and end of
each log file.  This includes the collect-support-data command-line
tool as well as the administrative task and extended operation that
can be used to invoke the tool.