Fix usage for multi-identifier args
Updated the ldapsearch, ldapmodify, ldapcompare, ldapdelete, and
manage-certificates tools to fix issues in which they exposed
alternative long identifiers for some arguments in the usage
information.  Although many arguments can be referenced by any of
multiple long identifiers, we generally only intend to expose one of
them in the usage output.

Also, fixed a typo in the ldifmodify usage output.