Add a test-ldap-sdk-performance tool
Added an internal command-line tool that can be used to help test
the performance of the LDAP SDK itself.  It creates an internal LDAP
listener that uses the canned response request handler to return a
predefined response for all search requests, and then invokes the
searchrate tool to issue searches as quickly as possible using a
specified number of threads.