Better control over passphrase-encrypted cipher
Updated the LDAP SDK's support for passphrase-based encryption to
provide better support for explicitly specifying the type of cipher
that should be used.  Previously, you could request either a
baseline level of protection (which would have always been based on
128-bit AES) or the strongest supported level of protection (which
would have either been based on 128-bit or 256-bit AES, depending on
what the underlying JVM supported).  It is now possible to
explicitly request 128-bit or 256-bit AES, with attempts to use the
latter failing if the JVM doesn't support it.  This framework will
provide even more flexibility if we add support for additional
cipher types in the future.