Support client auth in the LDAP listener
Updated the LDAP listener framework and the in-memory directory
server to provide support for mutual TLS authentication. in which
the server requests a certificate chain from the client and
validates any chain that is provided).  The server may be
configured to request that a client certificate chain, and it may
optionally choose to continue with or reject the TLS negotiation
attempt if the client does not present a certificate chain.