Add a parallel-update tool
Added a parallel-update command-line tool that can be used to apply
changes read from an LDIF file against an LDAP server using multiple
concurrent threads for higher throughput than single-threaded tools
like ldapmodify.  It properly handles dependencies between
concurrent operations (for example, it will delay adding a child
entry if the add for its parent is still in progress), and it also
supports retrying operations that fail in a way that indicates that
the operation might succeed if attempted after all other operations
have completed.  Any changes that cannot be processed successfully
will be written to a reject file for later examination.