Retry by default in ldapmodify and ldapdelete
Updated the ldapmodify and ldapdelete tools so that they will now
retry failed operations by default on a newly established connection
if the failure indicates that the underlying connection may no
longer be valid.  This behavior was previously available if the
--retryFailedOperations argument was provided, and that argument is
still available (but hidden) to preserve backward compatibility with
scripts that use it, but it now has no effect.  A new --neverRetry
argument has been added that can disable this feature if the retry
functionality should be disabled.