Fix a failing CachingNameResolver test
Updated the CachingNameResolverTestCase class to use www.github.com
instead of www.pingidentity.com for testing that depends on reverse
name resolution.  It seems that the www.pingidentity.com site has
been re-hosted and the IP address to which the hostname resolves
does not have a reverse record.  Since GitHub is used as the primary
repository for the LDAP SDK, and since the IP address for that
hostname does have a reverse record, the tests have been updated to
use that hostname and once again pass.