SSL socket verifier updates
Updated the HostNameSSLSocketVerifier to automatically trust any
certificate when the client connected to the peer using an IP
address that referenced a loopback interface on the system.  As per
the W3C Secure Contexts Candidate Recommendation
(https://www.w3.org/TR/secure-contexts/) section 3.2, connections
established with a loopback IP address should be considered
potentially trustworthy.

Updated the HostNameSSLSocketVerifier and TrustAllSSLSocketVerifier
classes to implement the javax.net.ssl.HostnameVerifier interface so
they may be used in contexts where that interface is accepted.