Close the command-line tool output file
Fixed an issue in the command-line tool framework in which the output
stream that the tool creates when writing output to a file (for tools
that support the outputFile argument, and when that argument is used
to send the output to a file) was never explicitly closed.  It was
implicitly closed when the tool exited, but this could cause problems
in cases where a program programmatically invoked the tool and then
tried to do something with the output file in the same JVM process.