Add manage-certificates retrieve-server-cert
Updated the manage-certificates tool to add a new
retrieve-server-certificate subcommand.  When this subcommand is
used, the tool will retrieve the certificate chain for a specified
remote server and write information about the certificates to
standard output.  It may also optionally write a PEM-encoded or
DER-encoded representation of that chain (or just the peer
certificate, if desired) to a specified file.