Provide programmatic access to an OID registry
Added a new OIDRegistry class that provides a registry of object
identifiers used by LDAP schema elements, controls, extended
operations, and other sources.  Each item has an OID, name, and
type, and it may optionally have an origin string and a URL to
content that can provide additional information about the item.

Added a new oid-lookup command-line tool that can be used to search
the OID registry to find items with a given OID, name, type, or
other content.