Add output format support to ldap-result-code
Updated the ldap-result-code tool to add support for an
--output-format argument that can cause the output to be formatted
as a user-friendly table (the default output), CSV, tab-delimited
text, or JSON objects.  The former --script-friendly argument (which
would cause the tool to format the output as tab-delimited text) has
been hidden in favor of the new --output-format argument, but is
still available and supported for the purpose of backward
compatibility.