Additional CryptoHelper methods
Updated the CryptoHelper to add a new getDefaultKeyStoreType method
that should be used instead of KeyStore.getDefaultType.

Also, added a CryptoHelper.setUseFIPSMode method that can be used to
programmatically enable FIPS 140-2-compliant mode.  It is still
recommended to enable FIPS mode by setting the
"com.unboundid.crypto.FIPS_MODE" system property to "true" before
the JVM is started, but this method can be used in cases where that
is not feasible.