Add a BC log level property
Added a new JVM property that can be used to control the log level
used for logging performed by the Bouncy Castle FIPS provider.  This
property will only be used if logging is enabled with a system
property, and its value must match the name of one of the
java.util.logging.Level constants.  If logging is enabled but the
log level property is not specified, then a default level of INFO
will be used.