Add an option to remove non-FIPS providers
Updated the crypto helper to make it possible to automatically
remove providers believed to be non-essential for operation in
FIPS-compliant mode and that may allow the use of non-approved
cryptographic operations.  This functionality is not enabled
by default, but can be used programmatically or enabled through the
com.unboundid.crypto.REMOVE_NON_ESSENTIAL_PROVIDERS property.