Add CryptoHelper methods for UUIDs
Updated the CryptoHelper class to add methods for obtaining random
and name-based UUIDs.  These methods should be used instead of the
methods that the UUID class provides, particularly when the LDAP SDK
is being used in FIPS 140-2-compliant mode:  when obtaining a
random UUID, the CryptoHelper implementation ensures that a
approved SecureRandom algorithm is used, and when obtaining a
name-based UUID, the CryptoHelper implementation uses the SHA-256
digest algorithm instead of the MD5 digest algorithm.