Add manage-certificates copy-keystore
Updated the manage-certificates tool to add a new copy-keystore
subcommand.  This subcommand can be used to copy all or some of the
certificates from one key store to another, creating the destination
key store if necessary.  The source and destination key stores can
be of different types, allowing you to convert one type of key store
to another (for example, converting a JKS key store to a PKCS #12
key store).